requirements/base.txt
aiohttp==2.3.7
amqp==2.2.2
anyjson==0.3.3
async-timeout==2.0.0
Babel==2.5.1
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
crontab==0.22.0
cryptography==1.7.2
Django==1.11.5
django-celery==3.2.2
django-cors-headers==2.1.0
django-debug-toolbar==1.8
django-environ==0.4.4
django-extensions==1.9.8
django-filter==1.1.0
django-redis==4.8.0
djangorestframework==3.7.3
flower==0.9.2
idna==2.6
inflection==0.3.1
jsonschema==2.6.0
kombu==4.1.0
Markdown==2.6.10
monotonic==1.4
multidict==3.3.2
psycopg2==2.7.3
pyasn1==0.4.2
pybreaker==0.4.2
pycparser==2.18
PyJWT==1.5.3
pytz==2017.3
raven==6.4.0
redis==2.10.6
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
tenacity==4.8.0
tornado==4.5.2
typing==3.6.2
urllib3==1.22
vine==1.1.4
yarl==0.17.0